chore: modify to server listen on index file
